# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APSO Backend API",
    "description": "AI-powered SDLC intelligence backend. MVP scope is read-only extraction and analysis.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/integrations/bitbucket/repositories": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Bitbucket Repositories",
        "operationId": "list_bitbucket_repositories_api_v1_integrations_bitbucket_repositories_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketRepositoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/{repo_slug}/pull-requests": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Bitbucket Pull Requests",
        "operationId": "list_bitbucket_pull_requests_api_v1_integrations_bitbucket_repositories__repo_slug__pull_requests_get",
        "parameters": [
          {
            "name": "repo_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Slug"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "OPEN",
              "title": "State"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketPullRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/{repo_slug}/commits": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Bitbucket Commits",
        "operationId": "list_bitbucket_commits_api_v1_integrations_bitbucket_repositories__repo_slug__commits_get",
        "parameters": [
          {
            "name": "repo_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Slug"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "develop",
              "title": "Branch"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketCommitListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/{repo_slug}/pipelines": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List Bitbucket Pipelines",
        "operationId": "list_bitbucket_pipelines_api_v1_integrations_bitbucket_repositories__repo_slug__pipelines_get",
        "parameters": [
          {
            "name": "repo_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketPipelineListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/ingest": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Ingest Bitbucket Repositories",
        "operationId": "ingest_bitbucket_repositories_api_v1_integrations_bitbucket_repositories_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BitbucketRepositoryIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketRepositoryIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/{repo_slug}/pull-requests/ingest": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Ingest Bitbucket Pull Requests",
        "operationId": "ingest_bitbucket_pull_requests_api_v1_integrations_bitbucket_repositories__repo_slug__pull_requests_ingest_post",
        "parameters": [
          {
            "name": "repo_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BitbucketPullRequestIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketPullRequestIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/bitbucket/repositories/{repo_slug}/commits/ingest": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Ingest Bitbucket Commits",
        "operationId": "ingest_bitbucket_commits_api_v1_integrations_bitbucket_repositories__repo_slug__commits_ingest_post",
        "parameters": [
          {
            "name": "repo_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BitbucketCommitIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BitbucketCommitIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BitbucketCommitListResponse": {
        "properties": {
          "commits": {
            "items": {
              "$ref": "#/components/schemas/BitbucketCommitSummary"
            },
            "type": "array",
            "title": "Commits"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "commits"
        ],
        "title": "BitbucketCommitListResponse"
      },
...
      "BitbucketPipelineListResponse": {
        "properties": {
          "pipelines": {
            "items": {
              "$ref": "#/components/schemas/BitbucketPipelineSummary"
            },
            "type": "array",
            "title": "Pipelines"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "pipelines"
        ],
        "title": "BitbucketPipelineListResponse"
      },
...
      "BitbucketPullRequestListResponse": {
        "properties": {
          "pull_requests": {
            "items": {
              "$ref": "#/components/schemas/BitbucketPullRequestSummary"
            },
            "type": "array",
            "title": "Pull Requests"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "pull_requests"
        ],
        "title": "BitbucketPullRequestListResponse"
      },
...
      "BitbucketRepositoryIngestRequest": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "title": "BitbucketRepositoryIngestRequest"
      },
      "BitbucketRepositoryIngestResponse": {
        "properties": {
          "received": {
            "type": "integer",
            "title": "Received"
          },
          "upserted": {
            "type": "integer",
            "title": "Upserted"
          },
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/BitbucketRepositorySummary"
            },
            "type": "array",
            "title": "Repositories"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "received",
          "upserted",
          "repositories"
        ],
        "title": "BitbucketRepositoryIngestResponse"
      },
      "BitbucketRepositoryListResponse": {
        "properties": {
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/BitbucketRepositorySummary"
            },
            "type": "array",
            "title": "Repositories"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "repositories"
        ],
        "title": "BitbucketRepositoryListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}